pyproject.toml
[tool.black]
line-length = 79

[tool.poetry]
name = "dns-assessment"
version = "0.1.0"
description = ""
authors = ["Rodrigo Martins <contact@martins.icu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-whois = "^0.8.0"
dnspython = "^2.4.2"
requests = "^2.31.0"
tqdm = "^4.66.1"
tldextract = "^5.1.1"
lxml = "^4.9.3"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pre-commit = "^3.6.0"
isort = "^5.13.1"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
